Develop

+++ src/channel/channel.service.ts
@@ -49,6 +49,12 @@ export class ChannelService {
   }
 
   async createSubscribe(userIdx: number, channelIdx: number): Promise<void> {
+    const subscribeState = await this.getSubscrbeState(userIdx, channelIdx);
+
+    if (subscribeState) {
+      throw new ConflictException('alreay subscribe');
+    }
+
     await this.prisma.subscribe.create({
       data: { subscriber: userIdx, provider: channelIdx },
     });
@@ -58,6 +64,12 @@ export class ChannelService {
     userIdx: number,
     channelIdx: number,
   ): Promise<void> {
+    const subscribeState = await this.getSubscrbeState(userIdx, channelIdx);
+
+    if (!subscribeState) {
+      throw new ConflictException('alreay subscribe');
+    }
+
     await this.prisma.subscribe.deleteMany({
       where: { subscriber: userIdx, provider: channelIdx },
     });
